Make sure ->app always holds a (weak) reference to the app
App introspection is possible by holding a reference to the app in each
of its commands. In the past, this was the package name of the app's
package, but this made it impossible to have eg. multiple instances of
an app, or to have an app whose state is not entirely known at compile
time.

This patch solves this by holding a reference instead, which is made
weak to reduce the risk of memory leaks. For consistency (and since
a common pattern seems to be for apps to subclass both App::CLI
and App::CLI::Command), the app also gets a weak reference to itself.

An additional change in cmd_map makes sure the package name is properly
constructed.