Make cmd_map map to possibly arbitrary packages
This patch modifies the behaviour of cmd_map to always return
the full name of the package implementing a given command, or
return undef trying.

Aliases that map to a string starting with a '+' sign are returned
as-is, mimicking the behaviour in other distributions. This makes
it possible to accept common commands (like 'help' or 'version') and
pass them to existing packages that can be extended by other users.