Add Usage class, outsourcing POD munging to Pod::Usage
Until now, the usage commands parsed the POD of a specified file
manually opening a filehandle and using Pod::Simple::Text. This was
done everywhere that required parsing the POD, which included the
default Help command, as well as the usage and brief_usage functions.

This commit implements a App::CLI::Usage class that is simply a wrapper
around Pod::Usage (or whatever other POD parser is passed as argument).