Remove abstract info from 'commands' command output
We now simply list the commands.  The reason for this change is that the
`abtract` method was using a non-existent class (the Usage class had
been added as part of a patch, however would have broken downstream
code) and the word 'abstract' is ambiguous as it looks like the keyword
of the same name, except all in capitals.  Here we also add a test for
the `commands` command to check that it does basically work as intended.